requirements.txt

pandas==2.2.1
requests==2.31.0
google-api-python-client==2.122.0
google-auth==2.28.2
openpyxl==3.1.2
xlwings==0.30.15
python-dotenv==1.0.1